% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{example_results}
\alias{example_results}
\alias{results}
\title{Results for example BAM-files}
\format{Files containing \code{BreakPoint} object.}
\source{
A lyphoblastoid cell line of Puerto Rican origin (HG00731). Raw reads were aligned to the GRCh38/hg38 reference assembly. For details see TODO:insert publication.
}
\description{
Localized breakpoints in example BAM-files, generated by the \pkg{breakpointR} package.
}
